requirements-dev.txt
-r requirements.txt
flake8==3.3.0
flake8-import-order==0.12
flake8-quotes==0.9.0
pytest==3.1.2
pytest-pep8==1.0.6
pytest-mock==1.5.0
pytest-catchlog==1.2.2
pytest-cov==2.4.0
yapf==0.16.1
PyYAML==3.12
attrdict==2.0.0
runscript==0.2.6
